pyproject.toml
[tool.poetry.dependencies]
python = ">=3.11,<3.15"

[project]
name = "human-face-emotions"
version = "0.1.0"
description = "Human emotions multi-classification using images of faces."
authors = [
    {name = "arseniyx92 khlytchiev.ad@phystech.edu"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "pytorch-lightning (>=2.6.0,<3.0.0)",
    "dvc (>=3.65.0,<4.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "torchmetrics (>=1.8.2,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "onnx (>=1.20.0,<2.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "onnxscript (>=0.5.7,<0.6.0)",
    "onnxruntime (>=1.23.2,<2.0.0)",
    "mlflow (>=3.8.1,<4.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)"
]
packages = [
    { include = "human_face_emotions", from="."}
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "fire (>=0.7.1,<0.8.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "nbqa (>=1.9.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
